rendering images in the filmstrip
