+++ renderer/filmstrip/filmstrip.css
@@ -1,0 +1,3 @@
+.filmstrip img {
+  height: 100%;
+}
