+++ renderer/image/image.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 
-const dcraw = require('dcraw');
+const { imageElem } = require('../util.js');
 
 let style = fs.readFileSync(path.resolve(__dirname, 'image.css'), 'utf8');
 
@@ -10,25 +10,7 @@ module.exports = function ({ events }) {
   elem.className = 'image';
 
   function loadImage(filepath) {
-    console.time('read');
-    const file = fs.readFileSync(filepath);
-    console.timeEnd('read');
-
-    console.time('meta');
-    var meta = dcraw(file, { verbose: true, identify: true });
-    console.timeEnd('meta');
-
-    console.log(meta);
-
-    console.time('tiff');
-    var tiff = dcraw(file, { extractThumbnail: true });
-  //  var tiff = dcraw(file, { exportAsTiff: true });
-    console.timeEnd('tiff');
-
-    console.log(tiff.length);
-
-    var img = document.createElement('img');
-    img.src = 'data:image/jpeg;base64,' + Buffer.from(tiff).toString('base64');
+    var img = imageElem(filepath);
 
     elem.innerHTML = '';
     elem.appendChild(img);
